Updates for ASTToken parsing (start AND end locations)
ASTNodes now have two SourceLocation objects- one for the start, one for
the end. Each one still represents a row/column. This new format relies
on updates to the PythonAST repo.

Additional work/testing might be required for interpolating the start
and end location(s) when that data was not able to be parsed from the
JSON object.